Clean up to remove <subpackage-s from import-control.xml
These appear to be left over from another project, or by gone days...
the current src/ code of this project (sevntu-checks) does not have any
such packages.

It's causing no real harm, but it's initially confusing to read the
import-control.xml when something goes wrong in it to see all these
useless declarations.

Signed-off-by: Michael Vorburger <mike@vorburger.ch>