minor: Convert \t TAB char to 4 spaces in sevntu-checkstyle-sonar-plugin
For all only 4 files which this project has, all of which still had
tabs, contrary to the main sevntu-check project code.

This will allow to activate the Checkstyle checks, the same as the
sevntu-check project has, in the pom.xml of
sevntu-checkstyle-sonar-plugin (it would currently fail because it,
correctly, enforces no tabs to be used).

This will help to see only real changes in pull requests which edit the
sevntu-checkstyle-sonar-plugin/.../checkstyle-extensions.xml file, which
every new check in sevntu-check must do.